Back out of @steven676 recent audio changes
I have confirmed that neither the aosp nor the pa builds are equipped to
work with these changes, whether with the prebuilt blobs or building from
source. I have not narrowed down which exact commit it is, but it is one
between the update recovery.fstab and current NookieDevs head.